simplify the code design
