+++ Frontend/src/components/ColorSelect/ColorSelect.js
@@ -1,3 +1,4 @@
+// ColorSelect.js
 import React, { useState, useEffect } from 'react';
 import './colorSelect.css';
 import ArrowIcon from './ArrowIcon';
@@ -7,17 +8,9 @@ const ColorSelect = ({ colors, onColorSelect, defaultValue }) => {
     const [selectedColor, setSelectedColor] = useState({});
 
     useEffect(() => {
-        if (defaultValue) {
-            const defaultColor = colors[defaultValue]; 
-            if (defaultColor) {
-                setSelectedColor({ key: defaultValue, value: defaultColor });
-            } else {
-                setSelectedColor({ key: Object.keys(colors)[0], value: colors[Object.keys(colors)[0]] }); 
-            }
-        } else {
-            const firstKey = Object.keys(colors)[0];
-            setSelectedColor({ key: firstKey, value: colors[firstKey] }); 
-        }
+        const initialKey = defaultValue || Object.keys(colors)[0];
+        const initialColor = colors[initialKey] || colors[Object.keys(colors)[0]];
+        setSelectedColor({ key: initialKey, value: initialColor });
     }, [colors, defaultValue]);
 
     const handleColorChange = (key) => {
@@ -28,14 +21,10 @@ const ColorSelect = ({ colors, onColorSelect, defaultValue }) => {
     };
 
     return (
-        <div style={{ position: 'relative', width: '100px' }}>
-            <div
-                className="selected-color"
-                onClick={() => setIsOpen(!isOpen)}
-                style={{display: 'flex',alignItems: 'center',justifyContent: 'space-between',borderRadius: '4px',cursor: 'pointer',padding: '5px',}}
-            >
+        <div className="color-select" onClick={() => setIsOpen(!isOpen)}>
+            <div className="selected-color">
                 <div className="color-circle" style={{ backgroundColor: selectedColor.value }} />
-                <ArrowIcon style={{ marginLeft: 'auto', color: '#ffffff' }} />
+                <ArrowIcon className="arrow-icon" />
             </div>
             {isOpen && (
                 <div className="dropdown">
@@ -44,10 +33,9 @@ const ColorSelect = ({ colors, onColorSelect, defaultValue }) => {
                             key={key}
                             onClick={() => handleColorChange(key)}
                             className="dropdown-option"
-                            style={{ display: 'flex', alignItems: 'center', padding: '10px', backgroundColor: '#0e0e27', cursor: 'pointer' }}
                         >
                             <div className="color-circle-option" style={{ backgroundColor: colors[key] }} />
-                            <span style={{ color: '#ffffff', fontSize: '11px', marginLeft: '10px' }}>{key}</span>
+                            <span className="option-label">{key}</span>
                         </div>
                     ))}
                 </div>
