+++ Frontend/src/components/ColorSelect/colorSelect.css
@@ -1,28 +1,63 @@
+/* colorSelect.css */
+.color-select {
+    position: relative;
+    width: 100px;
+}
+
+.selected-color {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border-radius: 4px;
+    cursor: pointer;
+    padding: 5px;
+}
+
+.arrow-icon {
+    margin-left: auto;
+    color: #ffffff;
+}
+
 .color-circle {
     width: 40px;
     height: 40px;
     border-radius: 50%;
     display: inline-block;
 }
+
 .color-circle-option {
     width: 20px;
     height: 20px;
     border-radius: 50%;
     display: inline-block;
 }
+
 .dropdown {
     position: absolute;
-    top: 110%; 
-    left: 50%; 
+    top: 110%;
+    left: 50%;
     transform: translateX(-50%);
     width: 150px;
-    border-radius:10px;
+    border-radius: 10px;
     overflow: hidden;
     z-index: 10;
     box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
 }
 
+.dropdown-option {
+    display: flex;
+    align-items: center;
+    padding: 10px;
+    background-color: #0e0e27;
+    cursor: pointer;
+}
 
 .dropdown-option:hover {
-    background-color: #0e0e27;
+    background-color: #1c1c3a; /* Slightly lighter on hover */
+}
+
+.option-label {
+    color: #ffffff;
+    font-size: 11px;
+    margin-left: 10px;
 }
