+++ Frontend/src/components/ControlTable/TableBodyContent.js
@@ -1,0 +1,65 @@
+// TableBodyContent.js
+import React from 'react';
+import StyledTableCell from './StyledTableCell';
+import { TableRow } from '@mui/material';
+import ColorSelect from '../ColorSelect/ColorSelect';
+
+const TableBodyContent = ({ rows, columns, columnColorsMap }) => (
+  <>
+    {rows.map((row) => (
+      <TableRow key={row.id}>
+        {columns.flatMap((column) => {
+          if (column.subColumns) {
+            return column.subColumns.map((subColumn) => {
+              const value = row[subColumn.field];
+              const columnType = subColumn.type;
+              const colorMap = columnColorsMap[columnType];
+
+              return (
+                <StyledTableCell key={subColumn.field} style={{ width: subColumn.width }} cellType="body">
+                  {colorMap && value && colorMap[value] !== undefined ? (
+                    <ColorSelect
+                      colors={colorMap}
+                      onColorSelect={(color) => console.log(`Selected ${color.key} for ${subColumn.field}`)}
+                      defaultValue={value}
+                    />
+                  ) : (
+                    value
+                  )}
+                </StyledTableCell>
+              );
+            });
+          } else {
+            if (column.renderCell) {
+              return (
+                <StyledTableCell key={column.field} cellType="body">
+                  {column.renderCell(row)}
+                </StyledTableCell>
+              );
+            } else {
+              const value = row[column.field];
+              const columnType = column.type;
+              const colorMap = columnColorsMap[columnType];
+
+              return (
+                <StyledTableCell key={column.field} style={{ width: column.width }} cellType="body">
+                  {colorMap && value && colorMap[value] !== undefined ? (
+                    <ColorSelect
+                      colors={colorMap}
+                      onColorSelect={(color) => console.log(`Selected ${color.key} for ${column.field}`)}
+                      defaultValue={value}
+                    />
+                  ) : (
+                    value
+                  )}
+                </StyledTableCell>
+              );
+            }
+          }
+        })}
+      </TableRow>
+    ))}
+  </>
+);
+
+export default TableBodyContent;
